dpdk: expose DPDKIfaceConfig through util-dpdk header file
